Crate forest_libp2p_bitswap

forest_libp2p_bitswap

Another libp2p bitswap(SPEC) implementation in Rust.

Features

Compatible with go-bitswap

Optional request manager

Prometheus metrics

Multiple async task API support, async-std and tokio(optional behind feature tokio)

Compiles into WebAssembly and works in browser. (examples/bitswap-in-browser)

Feature flags

tokio, disabled by default. Use task API(s) from tokio instead of async-std inside this crate.

Note: since async-std task API(s) are compatible with tokio runtime, it still works fine with tokio runtime when this feature is disabled. But it won’t work with async-std runtime if this feature is enabled.

Usage

Basic usage of BitswapBehaviour, for writing swarm event flow, sending or receiving a request or a response, checkout tests/go_compat.rs. Note that a request manager is needed for a real-world application.

use forest_libp2p_bitswap::BitswapBehaviour;

// Use default protocol ID(s), same with `go-bitswap` defaults
let behaviour = BitswapBehaviour::default();

// Use custom protocol ID(s)
let behaviour = BitswapBehaviour::new(&[b"/test/ipfs/bitswap/1.2.0"], Default::default());

To use the builtin request manager that is optimized for filecoin network, a data store that implements BitswapStoreRead and BitswapStoreReadWrite is required. For hooking request manager in swarm event flow, requesting a block via request manager API, checkout tests/request_manager.rs.

use forest_libp2p_bitswap::BitswapBehaviour;

let behaviour = BitswapBehaviour::default();
// Gets the associated request manager from the bitswap behaviour
// Note: The response is of type Arc<BitswapRequestManager> so that
// you can easily clone it, store it or send it around.
let bitswap_request_manager = behaviour.request_manager();

Modules

request_manager

Request manager implementation that is optimized for filecoin network usage

task

Re-exports API(s) from the chosen task library

Structs

BitswapBehaviour

A go-bitswap compatible protocol that is built on top of [request_response::Behaviour].

BitswapRequest

Bitswap request type

Enums

BitswapMessage

Bitswap message Enum type that is either a BitswapRequest or a BitswapResponse

BitswapResponse

Bitswap response type

RequestType

Type of a bitswap request

Traits

BitswapStoreRead

Trait implemented by a block store for reading.

BitswapStoreReadWrite

Trait implemented by a block store for reading and writing.

Functions

register_metrics

Register bitswap metrics

Type Definitions

BitswapBehaviourEvent

libp2p swarm network behavior event of bitswap